Initial commit. Basic lobby functionality!
Lets players join a lobby using the command /findlobby. Once in a lobby, the player will be assigned to be either a runner or a hunter mostly randomly except for a few exceptions to make sure the game works. Players can also use the command /votestart mt to vote to start the game without the max number of players.

TODO:
Create a config.yml so that we can change max players in a lobby etc. on the fly instead of having to recompile. Should be easy.

Implement the actual game lmao